Refactored some methods
