+++ packages/kokoas-server/src/api/docusign/contracts/construction/helpers/getPayMethodX.ts
@@ -1,0 +1,12 @@
+import { TContractData } from 'kokoas-server/src/api/kintone/getContractData';
+
+export const getPayMethodX = (
+  payMethod: TContractData['payMethod'],
+) =>{
+  switch (payMethod) {
+    case '持参': return 175;
+    case '集金': return 247;
+    case '振込': return 320;
+    default: throw new Error(`支払い方法が不明でした。${payMethod}`);
+  }
+};
