+++ packages/kokoas-server/src/api/docusign/contracts/construction/makeEnvelope.test.ts
@@ -7,7 +7,11 @@ it('should make envelop', async () => {
     projEstimateId: '74',
     userCode: 'RPA03',
   });
-  const result = await makeEnvelope(data, 'created', 'wetInk');
+  const result = await makeEnvelope({
+    data,
+    signMethod: 'wetInk',
+    status: 'created',
+  });
 
   expect(result).toMatchSnapshot();
 }, 30000);
